% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getFramePvals}
\alias{getFramePvals}
\title{Helper method which populates a vector with p-values for group difference in each biomarker from a single frame.}
\usage{
getFramePvals(dataFrame)
}
\arguments{
\item{dataFrame}{The data frame to generate p-values from.}
}
\value{
A vector containing the p-values for group difference for each biomarker in the specified data frame.  Note that if there are p biomarkers, this vector will be of length p.
}
\description{
This function tests for normality within groups, then applies the appropriate test for group difference.
}
