% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{isInputValid}
\alias{isInputValid}
\title{Helper method for the main method.}
\usage{
isInputValid(input)
}
\arguments{
\item{input}{A list() of the data frames passed as input to the main function.}
}
\value{
True if the input is valid, as defined above; false otherwise.
}
\description{
Tests to see if the input to the main function is valid.
Valid input is defined as 2-5 data frames, each with:
\itemize{
    \item The same number of columns.
    \item At least 2 unique values in the first column.
}
}
