% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{isNormByGroup}
\alias{isNormByGroup}
\title{Helper method which determines if a biomarker is normally distributed by group at alpha level 0.05.}
\usage{
isNormByGroup(dataFrame, marker)
}
\arguments{
\item{dataFrame}{The data frame to test for normality by group.}

\item{marker}{The biomarker number to be tested; e.g., for the nth biomarker, input integer n.
Note that marker n corresponds to column n+1 in the data frame.}
}
\value{
True if normality cannot be rejected for all groups; false otherwise.
}
\description{
This function will be called by the implementations of bullet 1 and 2 on page 7 of the project slides.
Note that the \code{shapiro.test()} function limits sample size to \code{5000}.
}
