% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pooledPValues}
\alias{pooledPValues}
\title{Pools the p-values between data frames using the specified method.}
\usage{
pooledPValues(method = "Fisher", ...)
}
\arguments{
\item{method}{The desired pooling method: "Fisher" for Fisher, "Stouffer" for Stouffer, "minP" for minimum p-value, or "maxP" for maximum p-value.}

\item{...}{The data frames to pool the p-values from.  Must be between 2 and 5 data frames, all with the same number of columns, and at least 2 groups per data frame.}
}
\value{
A vector containing pooled p-values for each biomarker.  For p biomarkers, the vector will be of length p.
}
\description{
Pools the p-values between data frames using the specified method.
}
